revert conftest
